TP_Final/requirements.txt
streamlit==1.39.0
requests==2.32.3
PyYAML==6.0.2
Pillow>=7.1.0,<11
numpy==1.26.4
opencv-python==4.11.0.86